Add spack.modules.error
